Refactor aggregator collections
- Pull collections out into a subfolder
- Eliminate unnecessary element arg from removal API
