+++ src/NRules/NRules/Aggregators/Collections/FactGrouping.cs
@@ -1,0 +1,14 @@
+using System.Linq;
+
+namespace NRules.Aggregators.Collections
+{
+    internal class FactGrouping<TKey, TElement> : FactCollection<TElement>, IGrouping<TKey, TElement>
+    {
+        public FactGrouping(TKey key)
+        {
+            Key = key;
+        }
+
+        public TKey Key { get; set; }
+    }
+}
