+++ src/NRules/NRules/Aggregators/SortedAggregator.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using NRules.Aggregators.Collections;
 using NRules.RuleModel;
 using NRules.Utilities;
